+++ src/App.jsx
@@ -1,34 +1,15 @@
 import React, { useState, useEffect } from "react";
-import { useDispatch } from "react-redux";
 import "./App.css";
-import { login, logout } from "./store/authSlice";
-import authService from "./appwrite/auth";
 import { Header, Footer } from "../src/components/index";
+import UsegetUser from "../src/components/index";
 
 function App() {
-  const [loading, setLoading] = useState(true);
-  const dispatch = useDispatch();
-  useEffect(() => {
-    authService
-      .getCurrentUser()
-      .then((userData) => {
-        if (userData) {
-          dispatch(login({ userData }));
-        } else {
-          dispatch(logout());
-        }
-      })
-      .catch((err) => console.log(err))
-      .finally(() => setLoading(false));
-  }, []);
-
+  const { loading } = UsegetUser();
   return !loading ? (
     <div className="min-h-screen flex flex-wrap content-between bg-gray-400">
       <div className="w-full block">
         <Header />
-        <main>
-          TODO: 
-        </main>
+        <main>TODO:</main>
         <Footer />
       </div>
     </div>
