+++ src/components/Logo.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-function Logo({ width = "100px" }) {
+function Logo({ width = "100px",  height ='100px' }) {
   return <div>Logo</div>;
 }
