+++ src/components/index.js
@@ -3,4 +3,5 @@ import LogoutBtn from "./Header/LogoutBtn";
 import Container from "./container/Container";
 import Logo from "./Logo";
 import Header from "./Header/Header";
-export { Footer, LogoutBtn, Container, Logo, Header };
+import UsegetUser from "./UsegetUser";
+export { Footer, LogoutBtn, Container, Logo, Header,UsegetUser };
